Reformatting, restructuring.
I think I fixed a bug in the logged colors. I reformatted a
bunch of code and restructured small parts of it to rely less
heavily on global variables. I moved some of the code into a
config file so that it wouldn't be so error prone to swap between
local development and production.